database: add database structure
This commit adds a database structure with prepared tables
in migrations. The tutorial on how to use the migrations and what
commands were used can be found in the docs/tutorials folder.
The models are generated too but kept blank. At the same time,
the database settings are added and a demo for checking
the connection with the database.